Colocate pyproject.toml metadata parsing with other file kinds
